Clean up inline-mode
