+++ pyrogram/client/methods/bots/answer_inline_query.py
@@ -16,22 +16,65 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with Pyrogram.  If not, see <http://www.gnu.org/licenses/>.
 
+from typing import List
+
 from pyrogram.api import functions, types
 from pyrogram.client.ext import BaseClient
+from ...types.inline_mode import InlineQueryResult
 
 
 class AnswerInlineQuery(BaseClient):
     def answer_inline_query(
         self,
         inline_query_id: str,
-        results: list,
+        results: List[InlineQueryResult],
         cache_time: int = 300,
         is_personal: bool = None,
         next_offset: str = "",
         switch_pm_text: str = "",
         switch_pm_parameter: str = ""
     ):
-        # TODO: Docs
+        """Use this method to send answers to an inline query.
+        No more than 50 results per query are allowed.
+
+        Args:
+            inline_query_id (``str``):
+                Unique identifier for the answered query.
+
+            results (List of :obj:`InlineQueryResult <pyrogram.InlineQueryResult>`):
+                A list of results for the inline query.
+
+            cache_time (``int``, *optional*):
+                The maximum amount of time in seconds that the result of the inline query may be cached on the server.
+                Defaults to 300.
+
+            is_personal (``bool``, *optional*):
+                Pass True, if results may be cached on the server side only for the user that sent the query.
+                By default, results may be returned to any user who sends the same query.
+
+            next_offset (``str``, *optional*):
+                Pass the offset that a client should send in the next query with the same text to receive more results.
+                Pass an empty string if there are no more results or if you don‘t support pagination.
+                Offset length can’t exceed 64 bytes.
+
+            switch_pm_text (``str``, *optional*):
+                If passed, clients will display a button with specified text that switches the user to a private chat
+                with the bot and sends the bot a start message with the parameter switch_pm_parameter
+
+            switch_pm_parameter (``str``, *optional*):
+                `Deep-linking <https://core.telegram.org/bots#deep-linking>`_ parameter for the /start message sent to
+                the bot when user presses the switch button. 1-64 characters, only A-Z, a-z, 0-9, _ and - are allowed.
+
+                Example: An inline bot that sends YouTube videos can ask the user to connect the bot to their YouTube
+                account to adapt search results accordingly. To do this, it displays a "Connect your YouTube account"
+                button above the results, or even before showing any. The user presses the button, switches to a private
+                chat with the bot and, in doing so, passes a start parameter that instructs the bot to return an oauth
+                link. Once done, the bot can offer a switch_inline button so that the user can easily return to the chat
+                where they wanted to use the bot's inline capabilities.
+
+        Returns:
+            On success, True is returned.
+        """
         return self.send(
             functions.messages.SetInlineBotResults(
                 query_id=int(inline_query_id),
