+++ pyrogram/client/types/inline_mode/__init__.py
@@ -16,13 +16,6 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with Pyrogram.  If not, see <http://www.gnu.org/licenses/>.
 
-from .callback_game import CallbackGame
-from .callback_query import CallbackQuery
-from .force_reply import ForceReply
-from .game_high_score import GameHighScore
-from .game_high_scores import GameHighScores
-from .inline_keyboard_button import InlineKeyboardButton
-from .inline_keyboard_markup import InlineKeyboardMarkup
-from .keyboard_button import KeyboardButton
-from .reply_keyboard_markup import ReplyKeyboardMarkup
-from .reply_keyboard_remove import ReplyKeyboardRemove
+from .inline_query import InlineQuery
+from .inline_query_result import InlineQueryResult
+from .inline_query_result_article import InlineQueryResultArticle
