+++ pyrogram/client/types/inline_mode/inline_query.py
@@ -20,15 +20,16 @@ from typing import List
 
 import pyrogram
 from pyrogram.api import types
-from ..bots.inline_query_result import InlineQueryResult
+from .inline_query_result import InlineQueryResult
 from ..messages_and_media import Location
 from ..pyrogram_type import PyrogramType
+from ..update import Update
 from ..user_and_chats import User
 
 
-class InlineQuery(PyrogramType):
+class InlineQuery(PyrogramType, Update):
     """This object represents an incoming inline query.
-    When the user sends an empty query, your bot could return some default or trending results
+    When the user sends an empty query, your bot could return some default or trending results.
 
     Args:
         id (``str``):
@@ -50,6 +51,7 @@ class InlineQuery(PyrogramType):
 
     def __init__(
         self,
+        *,
         client: "pyrogram.client.ext.BaseClient",
         id: str,
         from_user: User,
@@ -69,7 +71,6 @@ class InlineQuery(PyrogramType):
     @staticmethod
     def _parse(client, inline_query: types.UpdateBotInlineQuery, users: dict) -> "InlineQuery":
         return InlineQuery(
-            client=client,
             id=str(inline_query.query_id),
             from_user=User._parse(client, users[inline_query.user_id]),
             query=inline_query.query,
@@ -78,7 +79,8 @@ class InlineQuery(PyrogramType):
                 longitude=inline_query.geo.long,
                 latitude=inline_query.geo.lat,
                 client=client
-            ) if inline_query.geo else None
+            ) if inline_query.geo else None,
+            client=client
         )
 
     def answer(
@@ -90,6 +92,55 @@ class InlineQuery(PyrogramType):
         switch_pm_text: str = "",
         switch_pm_parameter: str = ""
     ):
+        """Bound method *answer* of :obj:`InlineQuery <pyrogram.InlineQuery>`.
+
+        Use this method as a shortcut for:
+
+        .. code-block:: python
+
+            client.answer_inline_query(
+                inline_query.id,
+                results=[...]
+            )
+
+        Example:
+            .. code-block:: python
+
+                inline_query.answer([...])
+
+        Args:
+            results (List of :obj:`InlineQueryResult <pyrogram.InlineQueryResult>`):
+                A list of results for the inline query.
+
+            cache_time (``int``, *optional*):
+                The maximum amount of time in seconds that the result of the inline query may be cached on the server.
+                Defaults to 300.
+
+            is_personal (``bool``, *optional*):
+                Pass True, if results may be cached on the server side only for the user that sent the query.
+                By default, results may be returned to any user who sends the same query.
+
+            next_offset (``str``, *optional*):
+                Pass the offset that a client should send in the next query with the same text to receive more results.
+                Pass an empty string if there are no more results or if you don‘t support pagination.
+                Offset length can’t exceed 64 bytes.
+
+            switch_pm_text (``str``, *optional*):
+                If passed, clients will display a button with specified text that switches the user to a private chat
+                with the bot and sends the bot a start message with the parameter switch_pm_parameter
+
+            switch_pm_parameter (``str``, *optional*):
+                `Deep-linking <https://core.telegram.org/bots#deep-linking>`_ parameter for the /start message sent to
+                the bot when user presses the switch button. 1-64 characters, only A-Z, a-z, 0-9, _ and - are allowed.
+
+                Example: An inline bot that sends YouTube videos can ask the user to connect the bot to their YouTube
+                account to adapt search results accordingly. To do this, it displays a "Connect your YouTube account"
+                button above the results, or even before showing any. The user presses the button, switches to a private
+                chat with the bot and, in doing so, passes a start parameter that instructs the bot to return an oauth
+                link. Once done, the bot can offer a switch_inline button so that the user can easily return to the chat
+                where they wanted to use the bot's inline capabilities.
+        """
+
         return self._client.answer_inline_query(
             inline_query_id=self.id,
             results=results,
