+++ pyrogram/client/types/inline_mode/todo/inline_query_result_audio.py
@@ -16,53 +16,56 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with Pyrogram.  If not, see <http://www.gnu.org/licenses/>.
 
-from pyrogram.api.core import Object
+from pyrogram.client.types.pyrogram_type import PyrogramType
 
 
-class InlineQueryResultCachedDocument(Object):
-    """Represents a link to a file stored on the Telegram servers. By default, this file will be sent by the user with an optional caption. Alternatively, you can use input_message_content to send a message with the specified content instead of the file.
+class InlineQueryResultAudio(PyrogramType):
+    """Represents a link to an mp3 audio file. By default, this audio file will be sent by the user. Alternatively, you can use input_message_content to send a message with the specified content instead of the audio.
 
     Attributes:
-        ID: ``0xb0700015``
+        ID: ``0xb0700004``
 
     Args:
         type (``str``):
-            Type of the result, must be document.
+            Type of the result, must be audio.
 
         id (``str``):
             Unique identifier for this result, 1-64 bytes.
 
+        audio_url (``str``):
+            A valid URL for the audio file.
+
         title (``str``):
-            Title for the result.
-
-        document_file_id (``str``):
-            A valid file identifier for the file.
-
-        description (``str``, optional):
-            Short description of the result.
+            Title.
 
         caption (``str``, optional):
-            Caption of the document to be sent, 0-200 characters.
+            Caption, 0-200 characters.
 
         parse_mode (``str``, optional):
             Send Markdown or HTML, if you want Telegram apps to show bold, italic, fixed-width text or inline URLs in the media caption.
+
+        performer (``str``, optional):
+            Performer.
+
+        audio_duration (``int`` ``32-bit``, optional):
+            Audio duration in seconds.
 
         reply_markup (:obj:`InlineKeyboardMarkup <pyrogram.types.InlineKeyboardMarkup>`, optional):
             Inline keyboard attached to the message.
 
         input_message_content (:obj:`InputMessageContent <pyrogram.types.InputMessageContent>`, optional):
-            Content of the message to be sent instead of the file.
+            Content of the message to be sent instead of the audio.
 
     """
-    ID = 0xb0700015
 
-    def __init__(self, type: str, id: str, title: str, document_file_id: str, description: str = None, caption: str = None, parse_mode: str = None, reply_markup=None, input_message_content=None):
+    def __init__(self, type: str, id: str, audio_url: str, title: str, caption: str = None, parse_mode: str = None, performer: str = None, audio_duration: int = None, reply_markup=None, input_message_content=None):
         self.type = type  # string
         self.id = id  # string
+        self.audio_url = audio_url  # string
         self.title = title  # string
-        self.document_file_id = document_file_id  # string
-        self.description = description  # flags.0?string
-        self.caption = caption  # flags.1?string
-        self.parse_mode = parse_mode  # flags.2?string
-        self.reply_markup = reply_markup  # flags.3?InlineKeyboardMarkup
-        self.input_message_content = input_message_content  # flags.4?InputMessageContent
+        self.caption = caption  # flags.0?string
+        self.parse_mode = parse_mode  # flags.1?string
+        self.performer = performer  # flags.2?string
+        self.audio_duration = audio_duration  # flags.3?int
+        self.reply_markup = reply_markup  # flags.4?InlineKeyboardMarkup
+        self.input_message_content = input_message_content  # flags.5?InputMessageContent
