+++ pyrogram/client/types/inline_mode/todo/inline_query_result_cached_document.py
@@ -16,33 +16,33 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with Pyrogram.  If not, see <http://www.gnu.org/licenses/>.
 
-from pyrogram.api.core import Object
+from pyrogram.client.types.pyrogram_type import PyrogramType
 
 
-class InlineQueryResultCachedPhoto(Object):
-    """Represents a link to a photo stored on the Telegram servers. By default, this photo will be sent by the user with an optional caption. Alternatively, you can use input_message_content to send a message with the specified content instead of the photo.
+class InlineQueryResultCachedDocument(PyrogramType):
+    """Represents a link to a file stored on the Telegram servers. By default, this file will be sent by the user with an optional caption. Alternatively, you can use input_message_content to send a message with the specified content instead of the file.
 
     Attributes:
-        ID: ``0xb0700011``
+        ID: ``0xb0700015``
 
     Args:
         type (``str``):
-            Type of the result, must be photo.
+            Type of the result, must be document.
 
         id (``str``):
             Unique identifier for this result, 1-64 bytes.
 
-        photo_file_id (``str``):
-            A valid file identifier of the photo.
-
-        title (``str``, optional):
+        title (``str``):
             Title for the result.
+
+        document_file_id (``str``):
+            A valid file identifier for the file.
 
         description (``str``, optional):
             Short description of the result.
 
         caption (``str``, optional):
-            Caption of the photo to be sent, 0-200 characters.
+            Caption of the document to be sent, 0-200 characters.
 
         parse_mode (``str``, optional):
             Send Markdown or HTML, if you want Telegram apps to show bold, italic, fixed-width text or inline URLs in the media caption.
@@ -51,18 +51,18 @@ class InlineQueryResultCachedPhoto(Object):
             Inline keyboard attached to the message.
 
         input_message_content (:obj:`InputMessageContent <pyrogram.types.InputMessageContent>`, optional):
-            Content of the message to be sent instead of the photo.
+            Content of the message to be sent instead of the file.
 
     """
-    ID = 0xb0700011
+    ID = 0xb0700015
 
-    def __init__(self, type: str, id: str, photo_file_id: str, title: str = None, description: str = None, caption: str = None, parse_mode: str = None, reply_markup=None, input_message_content=None):
+    def __init__(self, type: str, id: str, title: str, document_file_id: str, description: str = None, caption: str = None, parse_mode: str = None, reply_markup=None, input_message_content=None):
         self.type = type  # string
         self.id = id  # string
-        self.photo_file_id = photo_file_id  # string
-        self.title = title  # flags.0?string
-        self.description = description  # flags.1?string
-        self.caption = caption  # flags.2?string
-        self.parse_mode = parse_mode  # flags.3?string
-        self.reply_markup = reply_markup  # flags.4?InlineKeyboardMarkup
-        self.input_message_content = input_message_content  # flags.5?InputMessageContent
+        self.title = title  # string
+        self.document_file_id = document_file_id  # string
+        self.description = description  # flags.0?string
+        self.caption = caption  # flags.1?string
+        self.parse_mode = parse_mode  # flags.2?string
+        self.reply_markup = reply_markup  # flags.3?InlineKeyboardMarkup
+        self.input_message_content = input_message_content  # flags.4?InputMessageContent
