+++ pyrogram/client/types/inline_mode/todo/inline_query_result_cached_photo.py
@@ -16,57 +16,53 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with Pyrogram.  If not, see <http://www.gnu.org/licenses/>.
 
-from pyrogram.api.core import Object
+from pyrogram.client.types.pyrogram_type import PyrogramType
 
 
-class InlineQueryResultAudio(Object):
-    """Represents a link to an mp3 audio file. By default, this audio file will be sent by the user. Alternatively, you can use input_message_content to send a message with the specified content instead of the audio.
+class InlineQueryResultCachedPhoto(PyrogramType):
+    """Represents a link to a photo stored on the Telegram servers. By default, this photo will be sent by the user with an optional caption. Alternatively, you can use input_message_content to send a message with the specified content instead of the photo.
 
     Attributes:
-        ID: ``0xb0700004``
+        ID: ``0xb0700011``
 
     Args:
         type (``str``):
-            Type of the result, must be audio.
+            Type of the result, must be photo.
 
         id (``str``):
             Unique identifier for this result, 1-64 bytes.
 
-        audio_url (``str``):
-            A valid URL for the audio file.
+        photo_file_id (``str``):
+            A valid file identifier of the photo.
 
-        title (``str``):
-            Title.
+        title (``str``, optional):
+            Title for the result.
+
+        description (``str``, optional):
+            Short description of the result.
 
         caption (``str``, optional):
-            Caption, 0-200 characters.
+            Caption of the photo to be sent, 0-200 characters.
 
         parse_mode (``str``, optional):
             Send Markdown or HTML, if you want Telegram apps to show bold, italic, fixed-width text or inline URLs in the media caption.
-
-        performer (``str``, optional):
-            Performer.
-
-        audio_duration (``int`` ``32-bit``, optional):
-            Audio duration in seconds.
 
         reply_markup (:obj:`InlineKeyboardMarkup <pyrogram.types.InlineKeyboardMarkup>`, optional):
             Inline keyboard attached to the message.
 
         input_message_content (:obj:`InputMessageContent <pyrogram.types.InputMessageContent>`, optional):
-            Content of the message to be sent instead of the audio.
+            Content of the message to be sent instead of the photo.
 
     """
-    ID = 0xb0700004
+    ID = 0xb0700011
 
-    def __init__(self, type: str, id: str, audio_url: str, title: str, caption: str = None, parse_mode: str = None, performer: str = None, audio_duration: int = None, reply_markup=None, input_message_content=None):
+    def __init__(self, type: str, id: str, photo_file_id: str, title: str = None, description: str = None, caption: str = None, parse_mode: str = None, reply_markup=None, input_message_content=None):
         self.type = type  # string
         self.id = id  # string
-        self.audio_url = audio_url  # string
-        self.title = title  # string
-        self.caption = caption  # flags.0?string
-        self.parse_mode = parse_mode  # flags.1?string
-        self.performer = performer  # flags.2?string
-        self.audio_duration = audio_duration  # flags.3?int
+        self.photo_file_id = photo_file_id  # string
+        self.title = title  # flags.0?string
+        self.description = description  # flags.1?string
+        self.caption = caption  # flags.2?string
+        self.parse_mode = parse_mode  # flags.3?string
         self.reply_markup = reply_markup  # flags.4?InlineKeyboardMarkup
         self.input_message_content = input_message_content  # flags.5?InputMessageContent
