+++ pyrogram/client/types/inline_mode/todo/inline_query_result_venue.py
@@ -16,39 +16,45 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with Pyrogram.  If not, see <http://www.gnu.org/licenses/>.
 
-from pyrogram.api.core import Object
+from pyrogram.client.types.pyrogram_type import PyrogramType
 
 
-class InlineQueryResultContact(Object):
-    """Represents a contact with a phone number. By default, this contact will be sent by the user. Alternatively, you can use input_message_content to send a message with the specified content instead of the contact.
+class InlineQueryResultVenue(PyrogramType):
+    """Represents a venue. By default, the venue will be sent by the user. Alternatively, you can use input_message_content to send a message with the specified content instead of the venue.
 
     Attributes:
-        ID: ``0xb0700009``
+        ID: ``0xb0700008``
 
     Args:
         type (``str``):
-            Type of the result, must be contact.
+            Type of the result, must be venue.
 
         id (``str``):
             Unique identifier for this result, 1-64 Bytes.
 
-        phone_number (``str``):
-            Contact's phone number.
+        latitude (``float`` ``64-bit``):
+            Latitude of the venue location in degrees.
 
-        first_name (``str``):
-            Contact's first name.
+        longitude (``float`` ``64-bit``):
+            Longitude of the venue location in degrees.
 
-        last_name (``str``, optional):
-            Contact's last name.
+        title (``str``):
+            Title of the venue.
 
-        vcard (``str``, optional):
-            Additional data about the contact in the form of a vCard, 0-2048 bytes.
+        address (``str``):
+            Address of the venue.
+
+        foursquare_id (``str``, optional):
+            Foursquare identifier of the venue if known.
+
+        foursquare_type (``str``, optional):
+            Foursquare type of the venue, if known. (For example, "arts_entertainment/default", "arts_entertainment/aquarium" or "food/icecream".).
 
         reply_markup (:obj:`InlineKeyboardMarkup <pyrogram.types.InlineKeyboardMarkup>`, optional):
             Inline keyboard attached to the message.
 
         input_message_content (:obj:`InputMessageContent <pyrogram.types.InputMessageContent>`, optional):
-            Content of the message to be sent instead of the contact.
+            Content of the message to be sent instead of the venue.
 
         thumb_url (``str``, optional):
             Url of the thumbnail for the result.
@@ -60,15 +66,17 @@ class InlineQueryResultContact(Object):
             Thumbnail height.
 
     """
-    ID = 0xb0700009
+    ID = 0xb0700008
 
-    def __init__(self, type: str, id: str, phone_number: str, first_name: str, last_name: str = None, vcard: str = None, reply_markup=None, input_message_content=None, thumb_url: str = None, thumb_width: int = None, thumb_height: int = None):
+    def __init__(self, type: str, id: str, latitude: float, longitude: float, title: str, address: str, foursquare_id: str = None, foursquare_type: str = None, reply_markup=None, input_message_content=None, thumb_url: str = None, thumb_width: int = None, thumb_height: int = None):
         self.type = type  # string
         self.id = id  # string
-        self.phone_number = phone_number  # string
-        self.first_name = first_name  # string
-        self.last_name = last_name  # flags.0?string
-        self.vcard = vcard  # flags.1?string
+        self.latitude = latitude  # double
+        self.longitude = longitude  # double
+        self.title = title  # string
+        self.address = address  # string
+        self.foursquare_id = foursquare_id  # flags.0?string
+        self.foursquare_type = foursquare_type  # flags.1?string
         self.reply_markup = reply_markup  # flags.2?InlineKeyboardMarkup
         self.input_message_content = input_message_content  # flags.3?InputMessageContent
         self.thumb_url = thumb_url  # flags.4?string
