+++ pyrogram/client/types/input_message_content/input_text_message_content.py
@@ -17,20 +17,38 @@
 # along with Pyrogram.  If not, see <http://www.gnu.org/licenses/>.
 
 from pyrogram.api import types
-from pyrogram.client.style import HTML, Markdown
+from .input_message_content import InputMessageContent
+from ...style import HTML, Markdown
 
 
-class InputTextMessageContent:
+class InputTextMessageContent(InputMessageContent):
+    """This object represents the content of a text message to be sent as the result of an inline query.
+
+    Args:
+        message_text (``str``):
+            Text of the message to be sent, 1-4096 characters.
+
+        parse_mode (``str``, *optional*):
+            Use :obj:`MARKDOWN <pyrogram.ParseMode.MARKDOWN>` or :obj:`HTML <pyrogram.ParseMode.HTML>`
+            if you want Telegram apps to show bold, italic, fixed-width text or inline URLs in your message.
+            Defaults to Markdown.
+
+        disable_web_page_preview (``bool``, *optional*):
+            Disables link previews for links in this message.
+    """
+
+    __slots__ = ["message_text", "parse_mode", "disable_web_page_preview"]
+
     def __init__(self, message_text: str, parse_mode: str = "", disable_web_page_preview: bool = None):
+        super().__init__()
+
         self.message_text = message_text
         self.parse_mode = parse_mode
         self.disable_web_page_preview = disable_web_page_preview
-
-        self.style = HTML() if parse_mode.lower() == "html" else Markdown()
 
     def write(self, reply_markup):
         return types.InputBotInlineMessageText(
             no_webpage=self.disable_web_page_preview or None,
             reply_markup=reply_markup.write() if reply_markup else None,
-            **self.style.parse(self.message_text)
+            **(HTML() if self.parse_mode.lower() == "html" else Markdown()).parse(self.message_text)
         )
